FIX: Renames folder NameManager => NameManagerDialog
